[CXF-7571] Move method into CDI codebase.
